feat: Add player lives concept
Player now has an attribute for 'lives' which starts at 5 and
decreases by one every time the player encounters an enemy and resets.
Right now, there is no logic for lives = 0 but this needs to be
added. Additionally, updated the html and css so that the lives are
displayed to the user next to the score.