refactor: Continuous Enemies
Updates the enemy initialization to create 5 enemies instead of 30
enemies. Adds a Enemy.reset method to reset the enemy to off the left
of the board. Adds a if statement to Enemy.update to check if the
enemy has moved off the right side of the screen and reset the enemy
if true.