feat: Change player character when picking up kitty
When the player is in the same spot as a kitty, her character changes
to her 'holding' the kitty and the kitty sprite is placed off of the
grid so it is no longer visible. If she encounters an enemy in this
state, she resets to her original position and original player image
(with no kitty).