feat: Add kittyIdx variable to player and more
Updates a lot of comments in the app.js file. Adds a kittyIdx variable
to the player that corresponds to the index in allKitties of the currently
held kitten. Updates code to use this new function. Also checks to see
if a kitty is being held if the player walks off the edge of the screen
and resets the kitty if one is being held.